requirements.txt
langchain==0.0.197
chromadb==0.3.23
llama-cpp-python==0.1.50
urllib3==2.0.2
python-dotenv==1.0.0
